@@ -1,9 +1,9 @@
-using Newtonsoft.Json;
+using System.Text.Json.Serialization;
 
 namespace Azure.AISearch.FunctionApp.Models;
 
 public class SkillRequest
 {
-    [JsonProperty("values")]
+    [JsonPropertyName("values")]
     public IList<SkillRequestRecord> Values { get; set; } = new List<SkillRequestRecord>();
 }
